Change again totalMEm to size_t and add include I forgot last time. Stan tell me if it's working for you or not, because I cannot compile without size_t, probably due ti cuda 3.2
git-svn-id: https://icl.cs.utk.edu/svn/magma/branches/sc_release@240 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ src/auxiliary.cpp
@@ -10,6 +10,8 @@
 #include <cuda.h>
 #include <cuda_runtime_api.h>
 #include <cublas.h>
+#include <inttypes.h>
+
 #include "magma.h"
 
 /* ////////////////////////////////////////////////////////////////////////////
@@ -57,7 +59,7 @@ void printout_devices( )
   for( int idevice = 0; idevice < ndevices; idevice++ )
     {
       char name[200];
-      unsigned int totalMem;
+      size_t totalMem;
       int clock;
       CUdevice dev;
